requirements/requirements.txt
PySide6
PyQtWebEngine
opencv-python
numpy
pillow
manga-ocr
loguru
scikit-image
ultralytics
python-docx
requests
beautifulsoup4